Rename MFE power class
